+++ Core/Layer/Menus/MenuLayer.Render.cs
@@ -2,6 +2,8 @@ using Helion.Geometry;
 using Helion.Geometry.Boxes;
 using Helion.Geometry.Vectors;
 using Helion.Graphics;
+using Helion.Layer.Options;
+using Helion.Layer.Options.Sections;
 using Helion.Menus;
 using Helion.Menus.Base;
 using Helion.Menus.Base.Text;
@@ -11,8 +13,10 @@ using Helion.Render.Common.Enums;
 using Helion.Render.Common.Renderers;
 using Helion.Render.Common.Textures;
 using Helion.Util;
-using NAudio.Wave.SampleProviders;
+using Helion.Util.Extensions;
 using System;
+using System.Collections.Generic;
+using System.Text;
 using static Helion.Render.Common.RenderDimensions;
 
 namespace Helion.Layer.Menus;
@@ -22,6 +26,9 @@ public partial class MenuLayer
     private const int ActiveMillis = 500;
     private const int SelectedOffsetX = -32;
     private const int SelectedOffsetY = 5;
+
+    private readonly List<string> m_mapNameLines = [];
+    private readonly StringBuilder m_mapNameStringBuilder = new();
 
     private IMenuComponent? m_previousSelectedComponent;
     private IRenderableTextureHandle? m_saveGameTexture;
@@ -41,14 +48,17 @@ public partial class MenuLayer
         if (!m_menus.TryPeek(out Menu? menu))
             return;
 
-        bool shouldRenderSaveGameDetails =
-            menu is SaveMenu { IsSaveMenu: false }
-            && m_config.Game.ExtendedSaveGameInfo;
+        var saveMenu = menu.CurrentComponent is MenuSaveRowComponent;
+        var offsetY = menu.TopPixelPadding;
+        var detailsEnabled = m_config.Game.ExtendedSaveGameInfo;
+        var firstRow = true;
 
-        int offsetY = menu.TopPixelPadding;
+        if (saveMenu)
+            DrawSaveMenuBox(hud, detailsEnabled);
+
         for (int i = 0; i < menu.Components.Count; i++)
         {
-            IMenuComponent component = menu.Components[i];
+            var component = menu.Components[i];
             bool isSelected = ReferenceEquals(menu.CurrentComponent, component);
             bool wasSelected = ReferenceEquals(menu.CurrentComponent, m_previousSelectedComponent);
 
@@ -67,28 +77,35 @@ public partial class MenuLayer
                     DrawText(hud, largeTextComponent, ref offsetY);
                     break;
                 case MenuSaveRowComponent saveRowComponent:
-                    DrawSaveRow(hud, saveRowComponent, isSelected, wasSelected, ref offsetY, shouldRenderSaveGameDetails);
+                    if (firstRow)
+                    {
+                        offsetY++;
+                        firstRow = false;
+                    }
+                    DrawSaveRow(hud, (SaveMenu)menu, saveRowComponent, isSelected, wasSelected, ref offsetY, detailsEnabled);
                     break;
                 default:
                     throw new Exception($"Unexpected menu component type for drawing: {component.GetType().FullName}");
             }
 
             if (isSelected)
-            {
                 m_previousSelectedComponent = menu.CurrentComponent;
-            }
         }
 
-        if (menu.CurrentComponent is MenuSaveRowComponent && m_saveGameSummary != null)
-        {
+        if (saveMenu && m_saveGameSummary != null)
             RenderSaveGameDetails(hud);
-        }
     }
 
-    private void DrawText(IHudRenderContext hud, MenuTextComponent text, ref int offsetY)
+    private static void DrawText(IHudRenderContext hud, MenuTextComponent text, ref int offsetY)
     {
-        hud.Text(text.Text, text.FontName, text.Size, (0, offsetY), out Dimension area, both: Align.TopMiddle);
-        offsetY += area.Height;
+        var align = text.Align ?? Align.TopMiddle;
+        if (align != Align.TopMiddle)
+            offsetY = 0;
+
+        hud.Text(text.Text, text.FontName, text.Size, (0, offsetY), out Dimension area, both: align);
+
+        if (align == Align.TopMiddle)
+            offsetY += area.Height;
     }
 
     private void DrawImage(IHudRenderContext hud, MenuImageComponent image, bool isSelected, ref int offsetY, int upscalingFactor)
@@ -140,108 +157,106 @@ public partial class MenuLayer
         hud.Image(selectedName, drawPosition, both: image.ImageAlign);
     }
 
-    private void DrawSaveRow(IHudRenderContext hud, MenuSaveRowComponent saveRowComponent, bool isSelected,
+    private static int GetSaveRowWidth(bool detailsEnabled) => detailsEnabled ? 218 : 301;
+
+    private void DrawSaveRow(IHudRenderContext hud, SaveMenu saveMenu, MenuSaveRowComponent saveRowComponent, bool isSelected,
         bool wasPreviouslySelected, ref int offsetY, bool detailsEnabled)
     {
-        const int LeftOffset = 32;
-        const int DesiredRowVerticalPadding = 4;
-        const int SelectionOffsetX = 6;
         const string FontName = Constants.Fonts.Small;
         int fontSize = hud.GetFontMaxHeight(FontName);
-        const string LeftBarName = "M_LSLEFT";
-        const string MiddleBarName = "M_LSCNTR";
-        const string RightBarName = "M_LSRGHT";
+        int menuRowWidth = GetSaveRowWidth(detailsEnabled);
 
-        // draw row background
-        if (!hud.Textures.TryGet(LeftBarName, out var leftHandle) ||
-            !hud.Textures.TryGet(MiddleBarName, out var midHandle) ||
-            !hud.Textures.TryGet(RightBarName, out var rightHandle))
+        var textDimension = hud.MeasureText("_", FontName, fontSize);
+        var textHeight = textDimension.Height; 
+
+        string saveText;
+        if (isSelected && saveMenu.IsTypingName)
         {
-            return;
+            var rowWidth = menuRowWidth - 10;
+            //Account for cursor flashing
+            if (!saveRowComponent.Text.EndsWith('_'))
+                rowWidth -= textDimension.Width;
+            saveText = hud.GetTypedText(saveRowComponent.Text, FontName, fontSize, rowWidth - 4);
         }
-
-        int offsetX = LeftOffset;
-        Dimension leftDim = leftHandle.Dimension;
-        Dimension midDim = midHandle.Dimension;
-        Dimension rightDim = rightHandle.Dimension;
-
-        hud.Image(LeftBarName, (offsetX, offsetY));
-        offsetX += leftDim.Width;
-
-        int menuRowWidth = detailsEnabled ? 180 : 248;
-
-        int blocks = (int)Math.Ceiling((menuRowWidth - leftDim.Width - rightDim.Width) / (double)midDim.Width) + 1;
-        for (int i = 0; i < blocks; i++)
+        else
         {
-            hud.Image(MiddleBarName, (offsetX, offsetY));
-            offsetX += midDim.Width;
+            saveText = hud.GetEllipsesText(saveRowComponent.Text, FontName, fontSize, menuRowWidth - 4);
         }
+        
+        var rowHeight = textHeight + 3;
+        hud.PushOffset((17, 0));
 
-        hud.Image(RightBarName, (offsetX, offsetY));
-
-        int rowBgHeight = MathHelper.Max(leftDim.Height, midDim.Height, rightDim.Width);
-
-        // draw text
-        string saveText = saveRowComponent.Text.Length > blocks ? saveRowComponent.Text.Substring(0, blocks) : saveRowComponent.Text;
-        int textVerticalPadding = (rowBgHeight - fontSize) / 2;
-        Vec2I origin = (LeftOffset + leftDim.Width + 4, offsetY + textVerticalPadding);
-        hud.Text(saveText, FontName, fontSize, origin, out Dimension textArea);
-
-        // draw row selector
         if (isSelected)
         {
-            string selectedName = ShouldDrawActive ? Constants.MenuSelectIconActive : Constants.MenuSelectIconInactive;
-            if (hud.Textures.TryGet(selectedName, out var handle))
-            {
-                Vec2I selectedOffset = TranslateDoomOffset(handle.Offset);
-                selectedOffset += (LeftOffset - handle.Dimension.Width - SelectionOffsetX, offsetY - 2);
-                hud.Image(selectedName, selectedOffset);
-            }
+            hud.PushAlpha(0.5f);
+            HudBox box = new((0, offsetY), (menuRowWidth - 8, offsetY + rowHeight));
+            hud.FillBox(box, Color.Blue);
+            hud.PopAlpha();
         }
 
-        // Vanilla graphics are 14px high, some PWADs are higher. Target a 4px gap and eat into it if needed.
-        int rowTotalHeight = Math.Max(rowBgHeight, textArea.Height);
-        int rowVerticalPadding = Math.Max(0, (14 + DesiredRowVerticalPadding) - rowTotalHeight);
-        offsetY += rowTotalHeight + rowVerticalPadding;
+        hud.Text(saveText, FontName, fontSize, (1, offsetY + 2));
+        offsetY += rowHeight;
 
-        if (isSelected && detailsEnabled)
+        if (isSelected && detailsEnabled && !wasPreviouslySelected)
         {
-            if (!wasPreviouslySelected)
-            {
-                m_saveGameSummary = saveRowComponent.SaveGame == null
-                    ? null
-                    : new SaveGameSummary(saveRowComponent.SaveGame);
+            m_saveGameSummary = saveRowComponent.SaveGame == null
+                ? null
+                : new SaveGameSummary(saveRowComponent.SaveGame);
 
-                m_saveGameTexture = m_saveGameSummary?.UpdateSaveGameTexture(hud);
-            }
+            m_saveGameTexture = m_saveGameSummary?.UpdateSaveGameTexture(hud);
         }
+
+        hud.PopOffset();
+    }
+
+    private static void DrawSaveMenuBox(IHudRenderContext hud, bool detailsEnabled)
+    {
+        int height = 167;
+        hud.PushOffset((16, 20));
+        int saveRowWidth = GetSaveRowWidth(detailsEnabled);
+        var box = new HudBox((0, 0), (saveRowWidth - 6, height));
+        hud.PushAlpha(0.65f);
+        hud.BorderBox(box, Color.DarkGray, 1);
+        box = new HudBox((1, 1), (saveRowWidth - 7, height - 1));
+        hud.FillBox(box, Color.Black);
+        hud.PopAlpha();
+        hud.PopOffset();
     }
 
     private void RenderSaveGameDetails(IHudRenderContext hud)
     {
         if (m_saveGameSummary == null)
-        {
             return;
-        }
 
         const int TextSize = 4;
         const string Font = Constants.Fonts.Small;
         const int BoxWidth = 80;
         const int ThumbnailHeight = 60;
-        const int BoxHeight = ThumbnailHeight + 6 * 4 + 3;
+        int BoxHeight = ThumbnailHeight + 5 * TextSize + 3;
 
-        Vec2I boxUpperLeftBorder = (229, 31);
+        hud.LineWrap(m_saveGameSummary.MapName, Font, TextSize, BoxWidth - 4, m_mapNameLines, m_mapNameStringBuilder, 
+            out var requiredHeight);
+        BoxHeight += requiredHeight;
+
+        Vec2I boxUpperLeftBorder = (229, 20);
         Vec2I boxLowerRightBorder = boxUpperLeftBorder + (BoxWidth + 2, BoxHeight + 2);
 
-        Vec2I boxUpperLeft = (230, 32);
+        Vec2I boxUpperLeft = (230, 21);
         Vec2I boxLowerRight = boxUpperLeft + (BoxWidth, BoxHeight);
 
         hud.PushAlpha(0.65f);
-        DrawBorderBox(hud, new HudBox(boxUpperLeftBorder, boxLowerRightBorder), Color.DarkGray, 1);
+        hud.BorderBox(new HudBox(boxUpperLeftBorder, boxLowerRightBorder), Color.DarkGray, 1);
         hud.FillBox((boxUpperLeft, boxLowerRight), Color.Black);
         hud.PopAlpha();
 
-        if (m_saveGameTexture != null)
+        if (m_saveGameTexture == null)
+        {
+            hud.PushOffset(boxUpperLeft);
+            var size = hud.MeasureText("No Image", Font, TextSize);
+            hud.Text("No Image", Font, TextSize, (BoxWidth / 2 - size.Width / 2, ThumbnailHeight / 2 - size.Height / 2), textAlign: TextAlign.Center);
+            hud.PopOffset();
+        }
+        else
         {
             var imageBox = new HudBox(boxUpperLeft, boxUpperLeft + (BoxWidth, ThumbnailHeight));
             hud.Image(SaveGameSummary.TEXTURENAME, imageBox);
@@ -249,29 +264,20 @@ public partial class MenuLayer
 
         Vec2I offset = boxUpperLeft + (2, ThumbnailHeight + 2);
 
-        hud.Text(m_saveGameSummary.MapName, Font, TextSize, offset, out Dimension area, maxWidth: BoxWidth);
+        for (int i = 0; i< m_mapNameLines.Count; i++)
+        {
+            hud.Text(m_mapNameLines[i], Font, TextSize, offset, out var drawArea);
+            offset += (0, drawArea.Height);
+        }
+
+        hud.Text(m_saveGameSummary.Date, Font, TextSize, offset, out var area);
+        offset += (0, area.Height);
         offset += (0, area.Height);
 
-        hud.Text(m_saveGameSummary.Date, Font, TextSize, offset, out area);
-        offset += (0, area.Height);
-        offset += (0, area.Height);
-
-        foreach (string str in m_saveGameSummary?.Stats ?? Array.Empty<string>())
+        foreach (string str in m_saveGameSummary?.Stats ?? [])
         {
             hud.Text(str, Constants.Fonts.Small, 4, offset, out area);
             offset += (0, area.Height);
         }
-    }
-
-    private void DrawBorderBox(IHudRenderContext hud, HudBox box, Color color, int size)
-    {
-        HudBox topLine = new((box.TopLeft.X + size, box.TopLeft.Y), (box.TopRight.X - size, box.TopRight.Y + size));
-        HudBox bottomLine = new((box.BottomLeft.X + size, box.BottomLeft.Y - size), (box.BottomRight.X - size, box.BottomRight.Y));
-        HudBox leftLine = new(box.TopLeft, (box.BottomLeft.X + size, box.BottomLeft.Y));
-        HudBox rightLine = new((box.TopRight.X - size, box.TopRight.Y), box.BottomRight);
-        hud.FillBox(topLine, color);
-        hud.FillBox(bottomLine, color);
-        hud.FillBox(leftLine, color);
-        hud.FillBox(rightLine, color);
     }
 }