+++ Core/Layer/Options/Sections/ListedConfigSection.cs
@@ -641,7 +641,7 @@ public class ListedConfigSection : IOptionSection
                 name = header.HeaderText;
             }
 
-            name = GetEllipsesText(hud, name, Font, fontSize, hud.Dimension.Width / 2 - offsetX);
+            name = hud.GetEllipsesText(name, Font, fontSize, hud.Dimension.Width / 2 - offsetX);
             hud.Text(name, Font, fontSize, (-offsetX, y), out Dimension attrArea, window: Align.TopMiddle,
                 anchor: Align.TopRight, color: attrColor);
 
@@ -748,26 +748,6 @@ public class ListedConfigSection : IOptionSection
             anchor: Align.TopLeft);
         hud.FillBox((x + 1, y + 1, x + boxSize - 1, y + boxSize - 1), boxColor, window: Align.TopMiddle,
             anchor: Align.TopLeft);
-    }
-
-    public static string GetEllipsesText(IHudRenderContext hud, string text, string font, int fontSize, int maxWidth)
-    {
-        int nameWidth = hud.MeasureText(text, Font, fontSize).Width;
-        if (nameWidth <= maxWidth)
-            return text;
-
-        var textSpan = text.AsSpan();
-        int sub = 1;
-        while (sub < textSpan.Length && hud.MeasureText(textSpan, Font, fontSize).Width > maxWidth)
-        {
-            textSpan = text.AsSpan(0, text.Length - sub);
-            sub++;
-        }
-
-        if (textSpan.Length <= 3)
-            return text;
-
-        return string.Concat(text.AsSpan(0, textSpan.Length - 3), "...");
     }
 
     private bool IsConfigDisabled(int rowIndex)