+++ Core/Util/Extensions/HudExtensions.cs
@@ -4,6 +4,8 @@ using Helion.Render.Common.Renderers;
 using Helion.Render.Common.Textures;
 using Helion.Resources;
 using System;
+using System.Collections.Generic;
+using System.Text;
 
 namespace Helion.Util.Extensions;
 
@@ -49,5 +51,114 @@ public static class HudExtensions
     private static void HudVirtualStatusBar(HudImage hud)
     {
         hud.Hud.Image(hud.Image, (0, 0, hud.Handle.Dimension.Width, hud.Handle.Dimension.Height), both: Align.BottomLeft);
+    }
+
+    public static string GetEllipsesText(this IHudRenderContext hud, string text, string font, int fontSize, int maxWidth)
+    {
+        int nameWidth = hud.MeasureText(text, font, fontSize).Width;
+        if (nameWidth <= maxWidth)
+            return text;
+
+        var textSpan = text.AsSpan();
+        int sub = 1;
+        while (sub < textSpan.Length && hud.MeasureText(textSpan, font, fontSize).Width > maxWidth)
+        {
+            textSpan = text.AsSpan(0, text.Length - sub);
+            sub++;
+        }
+
+        if (textSpan.Length <= 3)
+            return text;
+
+        return string.Concat(text.AsSpan(0, textSpan.Length - 3), "...");
+    }
+
+    public static string GetTypedText(this IHudRenderContext hud, string text, string font, int fontSize, int maxWidth)
+    {
+        int nameWidth = hud.MeasureText(text, font, fontSize).Width;
+        if (nameWidth <= maxWidth)
+            return text;
+
+        var textSpan = text.AsSpan();
+        int sub = 1;
+        while (sub < textSpan.Length && hud.MeasureText(textSpan, font, fontSize).Width > maxWidth)
+        {
+            textSpan = text.AsSpan(sub, text.Length - sub);
+            sub++;
+        }
+
+        if (textSpan.Length - sub <= 0)
+            return string.Empty;
+
+        return text.AsSpan(sub).ToString();
+    }
+
+    public static ReadOnlySpan<char> TruncateText(this IHudRenderContext hud, string inputText, string font, int fontSize, int maxWidth)
+    {
+        if (string.IsNullOrEmpty(inputText))
+        {
+            return inputText;
+        }
+
+        for (int i = 0; i < inputText.Length; i++)
+        {
+            if (hud.MeasureText(inputText.AsSpan(0, i + 1), font, fontSize).Width > maxWidth)
+            {
+                return inputText.AsSpan(0, i);
+            }
+        }
+
+        return inputText;
+    }
+
+    public static void LineWrap(this IHudRenderContext hud, string inputText, string font, int fontSize, int maxWidth, List<string> lines, StringBuilder builder,
+        out int requiredHeight)
+    {
+        lines.Clear();
+        builder.Clear();
+        if (string.IsNullOrEmpty(inputText))
+        {
+            requiredHeight = 0;
+            return;
+        }
+
+        int maxTokenHeight = 0;
+        int widthCounter = 0;
+
+        int splitStart;
+        int splitEnd = 0;
+        for (int i = 0; i < inputText.Length; i++)
+        {
+            if (inputText[i] == ' ' || i == inputText.Length - 1)
+            {
+                splitStart = splitEnd;
+                splitEnd = i;
+            }
+            else
+            {
+                continue;
+            }
+
+            splitEnd++;
+            var token = inputText.AsSpan(splitStart, splitEnd - splitStart);
+            var tokenSize = hud.MeasureText(token, font, fontSize);
+            maxTokenHeight = Math.Max(maxTokenHeight, tokenSize.Height);
+
+            if (widthCounter + tokenSize.Width > maxWidth)
+            {
+                lines.Add(builder.ToString());
+                builder.Clear();
+                widthCounter = 0;
+            }
+
+            builder.Append(token);
+            widthCounter += tokenSize.Width;
+        }
+
+        // Flush the last line out of the StringBuilder
+        if (builder.Length > 0)
+            lines.Add(builder.ToString());
+
+        requiredHeight = lines.Count * maxTokenHeight;
     }
 }