+++ RELEASENOTES.md
@@ -20,7 +20,8 @@
   - Updated save game management to read save files on startup to fix load time of the load/save menu when there are a large number of save files to read
   - Improved quick save performance to reduce large stutters when saving
   - Add Boom RNG method with option to toggle between Boom and Vanilla Doom RNG
-  - Add save game thumbnails and summary details (kills, secrets, elapsed time) to Load menu
+  - Add save game thumbnails and summary details (kills, secrets, elapsed time) to save/load menus
+  - Updated styling of save/load menus
   - Add sky render modes dynamic and vanilla. Dynamic is the original version and similar to GZDoom. Vanilla to better match vanilla sky drawing
 
 ## Bug fixes: